00080-CASTOR-5279-Logical-Block-Protection-support-in-the-.patch 36acfd9a5b7a2818732586f70c85a29b7819ca15

+++ tapeserver/castor/tape/tapeserver/file/Structures.cpp
@@ -38,7 +38,7 @@ void tapeFile::VOL1::fill(std::string VSN,
   setString(m_ownerID, "CASTOR");
   std::stringstream hexLBP; 
   hexLBP << std::setfill('0') << std::setw(2) << std::hex
-      << std::noshowbase << LBPMethod;
+      << std::noshowbase << static_cast<int>(LBPMethod);
   setString(m_LBPMethod, hexLBP.str());
 }
